wip: refactor, running
