+++ agents/__init__.py
@@ -1,0 +1,3 @@
+from agents.legend import Legend
+from agents.npc import NPC
+from agents.saturnbot import SaturnBot
