+++ maze/models/maze.py
@@ -1,8 +1,6 @@
 import random
 import logging
 
-from autogen import (Agent, ConversableAgent, GroupChat, GroupChatManager,
-                     UserProxyAgent, config_list_from_json)
 from . import Activity, Cell, Item
 # Set up basic configuration for logging
 logging.basicConfig(
@@ -30,11 +28,22 @@ class Maze:
 
 
     def place_npcs(self, npcs):
-        # Place NPCs at specific or random locations as needed
+        # Place NPCs at specific and random locations as needed
+        status = False
         for npc in npcs:
-            x, y = (1, 1)  # Example fixed position, change as needed
-            self.maze_grid[x][y].place_npc(npc)
-            npc.send_initial_greeting()
+            # Place the first NPC in the starting location
+            if status == False:
+                self.place_npc(npc, self.start_point[0], self.start_point[1])
+                status = True
+            # Place the rest of the NPCs in random locations
+            else:
+                x, y = random.choice(self.get_all_locations())
+                self.place_npc(npc, x, y)
+            
+
+            # x, y = (1, 1)  # Example fixed position, change as needed
+            # self.maze_grid[x][y].place_npc(npc)
+            # npc.send_initial_greeting()
 
     def place_npc(self, npc, x, y):
         # Assigning NPC to the npc property of the cell
